backend/steam-auth/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
urllib3==2.0.7